+++ types/meson.build
@@ -3,6 +3,10 @@ wlr_files += files(
 	'data_device/wlr_data_offer.c',
 	'data_device/wlr_data_source.c',
 	'data_device/wlr_drag.c',
+	'output/cursor.c',
+	'output/output.c',
+	'output/render.c',
+	'output/transform.c',
 	'scene/subsurface_tree.c',
 	'scene/wlr_scene.c',
 	'scene/output_layout.c',
@@ -43,7 +47,6 @@ wlr_files += files(
 	'wlr_output_layout.c',
 	'wlr_output_management_v1.c',
 	'wlr_output_power_management_v1.c',
-	'wlr_output.c',
 	'wlr_pointer_constraints_v1.c',
 	'wlr_pointer_gestures_v1.c',
 	'wlr_pointer.c',
